[GH#2] Fix Starting an Unspecified Default Session
When the user has not specified a default session via the `user-session`
option in `lightdm.conf`, LightDM will return a session named `default`.
However, when asking LightDM to start a session, we must pass it as
`NULL` instead of the string `default`.

This change fixes a bug causing Authentication to seem broken, since
LightDM isn't able to find a `default.desktop` in
`/usr/share/xsessions`.

Fixes GH#2