[GH#51] Ensure the Mouse Cursor is Hidden
Add a new UI callback hide_mouse_cursor function to hide the cursor when
it is hovered over the background or main windows.

When initializing the UI, we move the mouse to the upper-left corner of
the primary monitor, to ensure it is hovering over the background window
and is hidden.

Closes GH#51