Refactor Configuration Parsing
Add a parse_greeter_string function to the config.c file for parsing a
string from a configuration group & option with a fallback value if the
option does not exist.

This reduces the amount of NULL checks we have to do in the
initialize_config function and should make it easier to provide default
values for every configuration option.