[#GH52] Add Option for Setting Password Input Width
Add a new `password-input-width` option to the `greeter` section of the
config file. This can be a negative number to defer to GTK's default
input width, or a positive number to suggest to GTK the number of
characters that should fit into the input.

Note: GTK may decide to render a narrower input.

Add a new parse_greeter_integer helper function for parsing integers
from the configuration file. A fallback value will be used if the option
is invalid or not present.

Closes GH#52