[#1687][GH#37] Disable the Password Input While Authenticating
Fix a bug causing LightDM to crash when the password was submitted while
a previous authentication was in progress. Disable the password input
when the password is submitted and wait until the authentication attempt
succeeds or fails to clear & re-enable it.

Add a new password_callback_id field to the App type to store the signal
handler ID for the handle_password callback on the password_input's
activate signal. This is populated when the app is initialized and the
initial signal handler is connected. When a password is submitted, we
use the field value to disconnect the handler from the signal. When
authentication succeeds or fails, we re-add the handler and store the
new handler ID in the field.

Closes #1687
Closes GH#37