[GH#68] Add Ability to Set Password Mask Character
Add a new `password-character` option to the `greeter-theme` section of
the configuration file. This allows users to modify the character that
is displayed in the input when they type their password.

A value of -1 means to use the default bullet, 0 means to display
nothing, & any other value sets the password mask character. Even emojis
work if your font supports them!

Closes GH#68