[GH#32] Add Ability to Switch Between DE/WM Sessions
A new `session-key` configuration option has been added. When pressed
with the `mod-key`, mini greeter will switch to the next session in
LightDM's session list & display it under the password input.

`user-session` no longer needs to be set in `lightdm.conf`, but we
pre-select it if it is a valid session.

Add a `FocusRing` type & related functions for keeping a selection
within a list & looping around at the ends.